+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## master
 
+
 - Use `send` instead of `public_send` to get the `authorization_context` so that contexts such as
   `current_user` can be `private` in the controller. ([@brendon][])
+
+- Fix railtie initialisation for Rails < 5. ([@brendon][])
+
 
 ## 0.2.0 (2018-06-17)
 
@@ -48,3 +52,4 @@
 
 [@palkan]: https://github.com/palkan
 [@ilyasgaraev]: https://github.com/ilyasgaraev
+[@brendon]: https://github.com/brendon